Added 64bit to project and disbled labels and set to 0 when no images found
There are issue with the 64bit compiled program. The compressed preview image does not show and it AVs on shutdown. I'd like to get the 64bit version to work so the script runner can access SSH/SCP commands without having to copy/move them outside of the default folder, that the 32bit program cannot access.