Removed code that was causing problems with 64bit version
There is a bug here, casting to a TListView without first checking it is a TListView.
There's currently no TlistViews in this project, so this can be safely removed.
It's interesting that the 32bit compiler was ok with this.

Also changed code so that the image and labels get cleared and reset when opening an invalid JPG using the file menu.