Thumbnail size is now configurable
Also added code to save the original file to the output directory (except when replacing originals) if no compression occurred or it resulted in a file size larger than the original.  Also fix issue whereby when saving new default setting, visited images that did not get Applied would not get the new default.